==> 00_make_file.R <==
# Execute this code to run all source code
rmarkdown::render("01_getting_data.Rmd")
rmarkdown::render("02_data_cleaning.Rmd")
rmarkdown::render("03_analysis.Rmd")
rmarkdown::render("10_paper.Rmd")
rmarkdown::render("11_poster_professional.Rmd")
rmarkdown::render("12_presentation_slides.Rmd")
