==> 02_data_cleaning.Rmd <==
---
title: "Data Cleaning"
author: "Evangeline Reynolds"
date: "2/2/2019"
output: html_document
---


The 2019 text was copied and pasted into a text document from: 

https://www.cnn.com/2019/02/05/politics/donald-trump-state-of-the-union-2019-transcript/index.html

```{r}
library(tidyverse)
trump_2019 <- paste0(readLines("raw_data/trump_2019_cnn.txt"), collapse = "")

sotu_texts <- read_csv(file = "raw_data/transcripts.csv") %>% 
  bind_rows(data_frame(date = lubridate::as_date("2019-02-05"), 
                       president = "Donald J. Trump", 
                       transcript = trump_2019))

# once 2019 sotu is given, then you can add this one before saving out the object

save(sotu_texts, file = "data_products/sotu_texts.Rdata")

```
